refactor CLI into modules, add rustdoc comments, add devnet smoke test
- Split 2100-line CLI monolith into 18 command modules under cli/src/commands/
- Add rustdoc comments to all public functions and structs in both programs
- Add devnet-smoke-test.sh script that exercises all presets and logs tx evidence
