+++ cli/src/commands/index.ts
@@ -1,0 +1,17 @@
+export { registerInitCommands } from "./init";
+export { registerMintCommands } from "./mint";
+export { registerBurnCommands } from "./burn";
+export { registerFreezeCommands } from "./freeze";
+export { registerStatusCommands } from "./status";
+export { registerSupplyCommands } from "./supply";
+export { registerBlacklistCommands } from "./blacklist";
+export { registerSeizeCommands } from "./seize";
+export { registerMintersCommands } from "./minters";
+export { registerRolesCommands } from "./roles";
+export { registerAllowlistCommands } from "./allowlist";
+export { registerAuthorityCommands } from "./authority";
+export { registerMetadataCommands } from "./metadata";
+export { registerOracleCommands } from "./oracle";
+export { registerTransferCommands } from "./transfer";
+export { registerHoldersCommands } from "./holders";
+export { registerAskCommands } from "./ask";
